Rename local-slots.ts to slot-types.ts
