+++ packages/sdk/src/models/decrypted-item.ts
@@ -13,7 +13,7 @@ import { IDEK } from '../services/service';
 import SlotHelpers from '../util/slot-helpers';
 import ItemMap from './item-map';
 import { ItemUpdate } from './item-update';
-import { NewSlot, SDKDecryptedSlot } from './local-slot';
+import { NewSlot, SDKDecryptedSlot } from './slot-types';
 
 /**
  * Wraps Items returned from the API that have been decrypted, usually by {@link ItemService}.
