+++ packages/sdk/src/models/new-item.ts
@@ -2,7 +2,7 @@ import { ItemTemplate, NestedSlotAttributes, PostItemsRequest, Slot } from '@mee
 import SlotHelpers from '../util/slot-helpers';
 import { slotToNewSlot } from '../util/transformers';
 import ItemChange from './item-change';
-import { NewSlot } from './local-slot';
+import { NewSlot } from './slot-types';
 import { SymmetricKey } from './symmetric-key';
 
 /** An Item which does not exist in the API */
