+++ packages/sdk/src/models/slot-types.ts
